feat: Add menu functionality to query.py
Implemented menu functionality to query.py by adding code that
continuously prompts the user until they choose the option to quit. The
code only executes options that are spelled correctly by the user and is
not case-sensitive. When the user enters a correct option, the code will
execute the respective desired function and then prompt the user again
with the menu.

issue #1