feat: Add searching by cast functionality
Implemented search functionality by taking an input phrase from the user
and searching the movies list for matches. Takes the search phrase and
movie cast member names and converts to lower case before comparing for
substring matches. The array that holds the cast member names that
matched is not reset when checking the cast of another movie.

issue #4